pyproject.toml
[project]
name = "python-project-52"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2.1",
    "django-admin>=2.0.2",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "load-dotenv>=0.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "rollbar>=1.3.0",
    "whitenoise>=6.9.0",
]
